requirements.txt
flask
python-dotenv
spotipy
notion-client
streamlit
pandas
pytest
